refactor: remove dead code an reorganize
- Remove unused older code
- Extract some logic from the view and put it inside services
- Remove tests that are not used anymore because too old
New tests should be written when possible
