+++ src/app/domain/osReceive/docs/nomenclature.md
